==> code/1.1_taxonomic_assignment_blastn.Rmd <==
---
title: "taxonomic assignment using blastn output"
author: "Kimberly Ledger"
date: "2024-11-08"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load-libraries}
library(tidyverse)
library(tidyr)
library(stringr)
rename <- dplyr::rename
```

## load output from taxonkit 
```{r load-output-from-taxonkit}
taxonomy <-read.delim("/home/kimberly.ledger/nearshore_eDNA/data/samples_w_mocks_updated/blastn_taxlineage.txt", header = FALSE, na.strings=c(""," ","NA"))

head(taxonomy)
```

## clean up the header a bit
```{r}
# use the full taxonomy rather than the seq id to collapse identical entries
tax_df <- taxonomy %>%
  filter(V4 > 100) %>% # make sure all retained matches are >100 bp
  select(-V2, -V5, -V6, -V7, -V14, -V8, -V9, -V10, -V11, -V12, -V13, -V15, -V16) %>%  #remove unnecessary columns
  group_by(V1, V17) %>% # group by the sequence key and the full taxonomy to reduce duplicate entries
  unique() %>% # doing that reduced the number of entries
  rename(qseqid=V1, perc_id=V3, length=V4, taxonomy=V17) %>% #rename headers
  #filter(perc_id > 98) %>% # blast output includes as low as 96 perc_id, but for now i only care about the top 2%
  filter(!is.na(taxonomy))
```

## calcualate the number of ASV's with taxonomic assignment
```{r how-many-total-ASVs-w-taxonomy}
tax_df %>% 
  ungroup() %>%
  select(qseqid) %>%
  unique() %>%
  nrow
```

## reformat table 
```{r}
# formatting the taxonomy variables
taxon_df <- tax_df %>%
  filter(str_detect(taxonomy, ";")) %>%
  separate(taxonomy, into=c("kingdom", "phylum", "class", "order", "family", "genus", "species"), sep = ";") %>% 
  filter(!grepl("sp\\.", species, ignore.case = TRUE)) %>%
  filter(!grepl("cf\\.", species, ignore.case = TRUE)) %>%
  filter(!grepl("aff\\.", species, ignore.case = TRUE)) %>%
  filter(!grepl(" x ", species, ignore.case = TRUE)) %>%         ### remove hybrids and unknown species assignments
  filter(!grepl("uncultured", species, ignore.case = TRUE)) %>%
  filter(!grepl("\\d", species))
```

## what non-fish are here? 
```{r}
not_Actinopteri <- taxon_df %>%
  filter(class != "Actinopteri") 

unique(not_Actinopteri$species)
```

## now look at the fish and figure out what fish taxa are in/out of our range
```{r}
to_check_range <- taxon_df %>%
  filter(class == "Actinopteri") %>%
  ungroup() %>%
  select(species, genus, family, order, class, phylum) %>%
  unique()
```

## check ranges for species using rfishbase 
```{r}
#library(remotes)
#remotes::install_github("ropensci/rfishbase")
library(rfishbase)

#one-time download of all fishbase tables... this takes a bit 
#fb_import()

#first, validate species names using rfishbase synonyms
spp_df <- synonyms(to_check_range$species)

syn <- spp_df %>% 
  filter(Status == "synonym")

#rename any synonyms with the validated name  
to_check_range_validated <- to_check_range %>%
  rename(synonym = species) %>%
  left_join(syn, by = "synonym") %>%
  mutate(synonym = ifelse(is.na(Species), synonym, Species)) %>%
  select(synonym:phylum) %>%
  rename(species = synonym)

#get distribution info 
spp_distribution <- distribution(to_check_range_validated$species) %>%
  select(Species, FAO) %>%
  unique()

#add column to designate if we will consider a species as "in range"- for this study, this will be NE Pacific and Arctic Ocean 
spp_distribution <- spp_distribution %>%
  mutate(in_range = ifelse(is.na(FAO), NA, "no"),
         in_range = ifelse(FAO == "Pacific, Northeast", "yes", in_range),
         in_range = ifelse(FAO == "America, North - Inland waters", "yes", in_range))

#keep just a list of spp names and yes/no/NA for "in range"  - this way we can keep track of what spp didn't have any reference information in fishbase to determine range 
spp_range <- spp_distribution %>%
  select(Species, in_range) %>%
  unique()

#how many entries do not have range info? 
range_na <- spp_range %>%
  filter(is.na(in_range))
```

When a valid name was not found, the presence of a species in the study area was checked using the GBIF database (https://www.gbif.org/).

## some species do not have range info - manually determine if these species should be considered in range
```{r}
spp_range <- spp_range %>%
  mutate(in_range = ifelse(Species == "Sebastes cheni", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Ammodytes japonicus", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Sebastiscus vibrantus", "no", in_range))

in_range <- spp_range %>%
  filter(in_range == "yes") 
```

## manually remove some species that do not belong in the dataset
```{r}
in_range <- in_range %>%
  filter(Species != "Oncorhynchus gilae") %>%
  filter(Species != "Acipenser brevirostrum") %>%  #positive control mis-id
  filter(Species != "Clupea harengus") %>%
  filter(Species != "Microgadus tomcod") %>%
  filter(Species != "Microcottus sellaris") %>%    ##in Bering Sea but not SE AK 
  filter(Species != "Oncorhynchus chrysogaster") %>%
  filter(Species != "Salvelinus leucomaenis") %>%
	filter(Species != "Salvelinus namaycush") %>%
  filter(!str_detect(Species, "Micropterus"))
```

## keep only the "in range" fish species in the taxon dataframe 
- this will remove the taxonomic hits for synonyms which i believe is fine
```{r}
taxon_df_reduced <- taxon_df %>%
  filter(species %in% in_range$Species)
```

## figure out what ASVs are unambiguously a single species
```{r}
single_spp_seqs <- taxon_df_reduced %>% 
  group_by(qseqid) %>%
  add_tally(name = "n_taxa") %>%
  filter(n_taxa == 1)
```

i quickly checked this species list and did not find any obvious species to remove 

## single species with at least 98% similarity 
```{r}
single_spp_seqs_98 <- single_spp_seqs %>%
  filter(perc_id >=98)
```

## remove the single-species seqs from the dataframe and then rank the hits by % identity for the remaining seqs
```{r}
seq_id_diff <- taxon_df_reduced %>%
  anti_join(., single_spp_seqs) %>%  ## comment out this line if there are no single species ASVs
  select(-length) %>%
  group_by(qseqid, species, genus, family, order, class, phylum, kingdom) %>%
    mutate(seq_percID = max(perc_id)) %>%
    group_by(qseqid, species, genus, family, order, class, phylum, kingdom, seq_percID) %>%
  summarise(max(seq_percID)) %>% # take just the top hit for each taxon (for each sequence)
  select(-`max(seq_percID)`) %>%
  ungroup() %>%
  group_by(qseqid) %>%
      #mutate(id_rank = rank(desc(seq_percID), ties.method = "min")) %>% # rank the taxonomic hits per sequence by % id
       mutate(top_perc = max(seq_percID)) %>% # designate the highest % id for the best taxonomic hit in each sequence (in some, but not all cases, this is 100%)   %>%
      mutate(diff = top_perc - seq_percID) %>% # calculate the difference between the % identity of the top hit and each subsequent taxonomic hit
        arrange(diff) %>%
  filter(top_perc >= 98)
```

Now you have the single best entry for each species for each sequence ranked and with the difference between the first and second ranked entries calculated.

For sequences with multiple top hits, where the difference between ranked taxa = 0, you will end up defaulting to genus- or family-level ID. We'll do the same for any sequences where the difference between the first and second ranked taxa is < 1%.

** if there is a sequence that is >99% match, then keep just the sequences within 0.5%  
** if there is NOT a >99% match, keep the top 1% 

## ASVs with >99% match or <99% match
```{r}
match_99 <- seq_id_diff %>%
  filter(top_perc >= 99) %>%
  select(qseqid) %>%
  unique()

match_lessthan99 <- seq_id_diff %>%
  filter(top_perc < 99) %>%
  select(qseqid) %>%
  unique()
```


## Figure out which differences are outside the threshold and eliminate those first
```{r}
# filter out any taxa that are >1% less matching identity than the top taxonomic hit for a given sequence
match_99_to_remove_low_perc_hits <- seq_id_diff %>%
  ungroup() %>%
  filter(qseqid %in% match_99$qseqid) %>%
  group_by(qseqid) %>%
  filter(diff > 0.5)

match_lessthan99_to_remove_low_perc_hits <- seq_id_diff %>%
  ungroup() %>%
  filter(qseqid %in% match_lessthan99$qseqid) %>%
  group_by(qseqid) %>%
  filter(diff > 1)

to_remove_low_perc_hits <- match_99_to_remove_low_perc_hits %>%
  bind_rows(match_lessthan99_to_remove_low_perc_hits)

keepers <- seq_id_diff %>%
  anti_join(to_remove_low_perc_hits)
```

## so now determine whether ASVs should be assigned to genus, family, order, etc. 
```{r}
singletons <- keepers %>%
  select(qseqid) %>%
  tally() %>%
  filter(n == 1)

# these are the seqs that now have only a single match
singleton_df <- singletons %>%
  left_join(keepers) %>%
  select(-n) %>%
  bind_rows(single_spp_seqs_98) %>% # combine the single spp data
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species)

## Genus-level matches
# remove the singletons from the bigger df 
single_genus <- keepers %>%
  anti_join(singleton_df)%>% # at best, these should be genus-level matches
  group_by(qseqid, genus) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single genus
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "genus") %>%
  mutate(taxon = genus)

## Family-level matches
single_family <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  group_by(qseqid, family) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single family
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "family") %>%
  mutate(taxon = family)

## Order-level matches
single_order <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  group_by(qseqid, order) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single order
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "order") %>%
  mutate(taxon = order)

## Class-level matches
single_class <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  anti_join(single_order) %>%
  group_by(qseqid, class) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>% 
  filter(n == 1) %>% # seqs that match a single class
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "class") %>%
  mutate(taxon = class)

## Phylum-level matches
single_phylum <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  anti_join(single_order) %>%
  anti_join(single_class) %>%
  group_by(qseqid, phylum) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>% 
  filter(n == 1) %>% # seqs that match a single phylum
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "phylum") %>%
  mutate(taxon = phylum)

# Kingdom level matches
single_kingdom <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  anti_join(single_order) %>%
  anti_join(single_class) %>%
  group_by(qseqid, kingdom) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>% 
  filter(n == 1) %>% # seqs that match a single kingdom
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "kingdom") %>%
  mutate(taxon = kingdom)
```

## modify the singleton_df to include the right variable headers
```{r}
single_spp <- singleton_df %>%
  select(-perc_id, -length, -n_taxa) %>%
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species)
```

## recombine the full data set now that the appropriate level of taxonomy has been determined
```{r}
tax_df_2 <- bind_rows(single_kingdom, single_phylum, single_class, single_order, single_family, single_genus, single_spp)
```

## also ran blastn against the salmon specific reference database 
blastn -query /home/kimberly.ledger/nearshore_eDNA/data/samples_w_mocks_updated/filtered.seqs.fasta -db /genetics/edna/custom_dbs/salmon_mifish_db_20241227 -out home/kimberly.ledger/nearshore_eDNA/data/samples_w_mocks_updated/blastn_salmon.txt -perc_identity 96 -qcov_hsp_perc 98 -num_threads 10 -outfmt '6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore sscinames staxids'

```{r}
salmon_tax <- read.delim("/home/kimberly.ledger/nearshore_eDNA/data/samples_w_mocks_updated/blastn_salmon.txt", header = FALSE, na.strings=c(""," ","NA"))

salmon_tax <- salmon_tax %>%
  filter(V4 > 100) %>% # make sure all retained matches are >100 bp
  select(-V5, -V6, -V7, -V8, -V9, -V10, -V11, -V12, -V13, -V14) %>%  #remove unnecessary columns
  group_by(V1, V2) %>% # group by the sequence key and the full taxonomy to reduce duplicate entries
  unique() %>% # doing that reduced the number of entries
  rename(qseqid=V1, species=V2, perc_id=V3, length=V4) #rename headers
```

## rank the hits by % identity
```{r}
seq_id_diff_salmon <- salmon_tax %>%
  select(-length) %>%
  group_by(qseqid, species) %>%
    mutate(seq_percID = max(perc_id)) %>%
    group_by(qseqid, species, seq_percID) %>%
  summarise(max(seq_percID)) %>% # take just the top hit for each taxon (for each sequence)
  select(-`max(seq_percID)`) %>%
  ungroup() %>%
  group_by(qseqid) %>%
      #mutate(id_rank = rank(desc(seq_percID), ties.method = "min")) %>% # rank the taxonomic hits per sequence by % id
       mutate(top_perc = max(seq_percID)) %>% # designate the highest % id for the best taxonomic hit in each sequence (in some, but not all cases, this is 100%)   %>%
      mutate(diff = top_perc - seq_percID) %>% # calculate the difference between the % identity of the top hit and each subsequent taxonomic hit
      arrange(diff)

seq_id_diff_salmon
```

** if there is a sequence that is >99% match, then keep just the sequences within 0.5%  
** if there is NOT a >99% match, keep the top 1% 

## Figure out which differences are outside the threshold and eliminate those first
```{r}
# filter out any taxa that are >1% less matching identity than the top taxonomic hit for a given sequence
to_remove_low_perc_hits_salmon <- seq_id_diff_salmon %>%
  ungroup() %>%
  group_by(qseqid) %>%
  filter(diff > 0.5)

keepers_salmon <- seq_id_diff_salmon %>%
  anti_join(to_remove_low_perc_hits_salmon) %>%
  filter(top_perc > 98)
```

## so now determine whether ASVs should be assigned to genus, family, order, etc. 
```{r}
singletons_salmon <- keepers_salmon %>%
  select(qseqid) %>%
  tally() %>%
  filter(n == 1)

# these are the seqs that now have only a single match
singleton_df_salmon <- singletons_salmon %>%
  left_join(keepers_salmon) %>%
  select(-n) %>%
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species)

## Genus-level matches
# remove the singletons from the bigger df 
single_genus_salmon <- keepers_salmon %>%
  anti_join(singleton_df_salmon)%>% # at best, these should be genus-level matches
  left_join(., keepers_salmon) %>%
  arrange(qseqid) %>%
  mutate(taxonomic_level = "genus") %>%
  mutate(taxon = "Oncorhynchus")
```

remove underscore from taxon name 
```{r}
salmon_df <- singleton_df_salmon %>%
  bind_rows(single_genus_salmon)

salmon_df$taxon <- gsub("_", " ", salmon_df$taxon)
salmon_df$species <- gsub("_", " ", salmon_df$species)
salmon_df$genus <- "Oncorhynchus"
salmon_df$family <- "Salmonidae"
salmon_df$order <- "Salmoniformes"
salmon_df$class <- "Actinopteri"
salmon_df$phylum <- "Chordata"
salmon_df$kingdom <- "Eukaryota"
```

## compare original salmon assignments to custom db assignments 
```{r}
salmon_comparison <- tax_df_2 %>%
  filter(genus == "Oncorhynchus") %>%
  select(qseqid, species, seq_percID, taxon) %>%
  left_join(salmon_df, by = "qseqid")
```
only a few differences between full ncbi db and custom salmon db. but will go ahead with the custom db verison. 

```{r}
tax_df_3 <- tax_df_2 %>%
  filter(genus != "Oncorhynchus") %>%
  bind_rows(salmon_df)
```

## custom updates to make: 
ASV_0053 - filter by diff = 0 to assign to Ophiodon elongatus
assign Pleuronectidae into groups 
Pleuronectidae 1: Hippoglossoides elassodon, Hippoglossoides robustus, Limanda aspera
Pleuronectidae 2: Isopsetta isolepis, Parophrys vetulus, Psettichthys melanostictus
Pleuronectidae 3: Liopsetta glacialis, Platichthys stellatus
Sebastes 1: Sebastes caurinus/Sebastes maliger
Sebastes 2: Sebastes auriculatus/Sebastes rastrelliger
Sebastes 3: lots of Sebastes sp including Sebastes flavidus
-put all Gadus reads to genus level. 
- put all Ammodytes reads to Ammodytes personatus (it's the only species present in region)
- lump the few Anoplarchus purpurescens reads to genus level 
- lump the few Cottus cognatus reads to genus level
- lump the few Myoxocephalus stelleri to genus level
- put Pholis reads to Pholis laeta
- move Sebastes nigrocinctus and Sebastes entomelas reads to Sebastes 3 ***


```{r}
tax_df_updated <- tax_df_3 %>%
  mutate(taxon = ifelse(qseqid %in% c("ASV_0053"), "Ophiodon elongatus", taxon), taxonomic_level = ifelse(qseqid %in% c("ASV_0053"), "species", taxonomic_level),
       taxon = ifelse(qseqid %in% c("ASV_0029"), "Pleuronectidae 1", taxon),
       taxon = ifelse(qseqid %in% c("ASV_0038", "ASV_0420", "ASV_0431", "ASV_0470", "ASV_0496", "ASV_0554", "ASV_0622", "ASV_0623"), "Pleuronectidae 2", taxon),
       taxon = ifelse(qseqid %in% c("ASV_0071", "ASV_0175", "ASV_0365"), "Pleuronectidae 3", taxon),
       taxon = ifelse(qseqid %in% c("ASV_0019", "ASV_0176", "ASV_0180", "ASV_0356", "ASV_0598"), "Sebastes 1", taxon),
       taxon = ifelse(qseqid %in% c("ASV_0023", "ASV_0168", "ASV_0332"), "Sebastes 2", taxon),
       taxon = ifelse(qseqid %in% c("ASV_0027","ASV_0125","ASV_0128", "ASV_0330", "ASV_0190"), "Sebastes 3", taxon), taxonomic_level = ifelse(genus == "Sebates 3", "genus", taxonomic_level),
       taxon = ifelse(genus == "Gadus", "Gadus", taxon), taxonomic_level = ifelse(genus == "Gadus", "genus", taxonomic_level),
       taxon = ifelse(qseqid %in% c("ASV_0020", "ASV_0044"), "Oncorhynchus keta", taxon),
       taxon = ifelse(qseqid %in% c("ASV_0083"), "Oncorhynchus kisutch", taxon),
       taxonomic_level = ifelse(qseqid %in% c("ASV_0020", "ASV_0044","ASV_0083"), "species", taxonomic_level),
       taxon = ifelse(qseqid %in% c("ASV_0108"), "Ammodytes personatus", taxon),
       taxonomic_level = ifelse(qseqid %in% c("ASV_0108"), "species", taxonomic_level), 
       taxon = ifelse(qseqid %in% c("ASV_0262"), "Anoplarchus", taxon),
       taxonomic_level = ifelse(qseqid %in% c("ASV_0262"), "genus", taxonomic_level),
       taxon = ifelse(qseqid %in% c("ASV_0608"), "Cottus", taxon),
       taxonomic_level = ifelse(qseqid %in% c("ASV_0608"), "genus", taxonomic_level),
       taxon = ifelse(qseqid %in% c("ASV_0272"), "Myoxocephalus", taxon),
       taxonomic_level = ifelse(qseqid %in% c("ASV_0272"), "genus", taxonomic_level),
       taxon = ifelse(qseqid %in% c("ASV_0236"), "Pholis laeta", taxon),
       taxonomic_level = ifelse(qseqid %in% c("ASV_0236"), "species", taxonomic_level), 
       taxonomic_level = ifelse(qseqid %in% c("ASV_0227", "ASV_0571", "ASV_0636", "ASV_0814"), "genus", taxonomic_level),
       taxon = ifelse(qseqid %in% c("ASV_0227", "ASV_0571", "ASV_0636", "ASV_0814"), "Cottus_Leptocottus", taxon)) %>%
    filter(qseqid != "ASV_0610") %>% # this asv only has couple reads in mock community samples (will get removed anyways...)
    filter(qseqid != "ASV_0864") %>%  #this asv only has 1 reads in mock community samples (will get removed anyways...)
    filter(qseqid != "ASV_0865")  #this asv only has 1 reads in mock community samples (will get removed anyways...)
```


```{r}
# and then collapse that down to just a single taxon per ASV
tax_df_updated_collapsed <- tax_df_updated %>%
  mutate(species = ifelse(taxonomic_level == "family", NA, species),
         genus = ifelse(taxonomic_level == "family", NA, genus),
         species = ifelse(taxonomic_level == "genus", NA, species)) %>%
  select(qseqid, taxon, taxonomic_level, species, genus, family, order, class, phylum, kingdom) %>%
  unique()
```

```{r}
write.csv(tax_df_updated, "/home/kimberly.ledger/nearshore_eDNA/outputs/samples_w_mocks_updated/taxonomy_20250110_uncollapsed.csv")
write.csv(tax_df_updated_collapsed, "/home/kimberly.ledger/nearshore_eDNA/outputs/samples_w_mocks_updated/taxonomy_20250110_collapsed.csv")
```


## output table of taxonomic groupings - need to make some manual edits to reflect customized changes above 

```{r}
groups_to_remove_1 <- tax_df_updated %>%
  filter(species == "Hexagrammos lagocephalus" | species == "Hexagrammos octogrammus" | species == "Hexagrammos stelleri") %>%
  filter(taxon == "Ophiodon elongatus")

groups_to_remove_2 <- tax_df_updated %>%
  filter(taxon == "Pleuronectidae 3") %>%
  filter(species == "Hippoglossus stenolepis")
  
groups_to_remove_3 <- tax_df_updated %>%
  filter(taxon == "Oncorhynchus kisutch" & species == "Oncorhynchus keta")

groups_to_remove_4 <- tax_df_updated %>%
  filter(taxon == "Oncorhynchus keta" & species == "Oncorhynchus kisutch")

groups_to_remove_5 <- tax_df_updated %>%
  filter(taxon == "Oncorhynchus keta" & species == "Oncorhynchus nerka")

groups_to_remove_6 <- tax_df_updated %>%
  filter(taxon == "Pholis laeta" & species == "Pholis fasciata")

groups_to_remove_7 <- tax_df_updated %>%
  filter(taxon == "Pholis laeta" & species == "Pholis clemensi")

groups_to_remove_8 <- tax_df_updated %>%
  filter(species == "Sebastes auriculatus" | species == "Sebastes rastrelliger") %>%
  filter(qseqid == "ASV_0598")

tax_df_groups <- tax_df_updated %>%
  #filter(taxonomic_level != "species") %>% 
  filter(species != "Ammodytes hexapterus") %>%
  anti_join(groups_to_remove_1) %>%
  anti_join(groups_to_remove_2) %>%
  anti_join(groups_to_remove_3) %>%
  anti_join(groups_to_remove_4) %>%
  anti_join(groups_to_remove_5) %>%
  anti_join(groups_to_remove_6) %>%
  anti_join(groups_to_remove_7) %>%
  anti_join(groups_to_remove_8) %>%
  select(taxon, taxonomic_level, species, genus, family, order) %>%
  unique() %>%
  arrange(taxon)
```

export 
```{r}
write.csv(tax_df_groups, "/home/kimberly.ledger/nearshore_eDNA/outputs/samples_w_mocks_updated/taxonomy_groups_20250110.csv")
```

Oncorhynchus = mykiss, nerka, keta, gorbuscha, kisutch
