==> code/7_qm_field_plots.Rmd <==
---
title: "qm field samples"
output: html_document
date: "2024-11-25"
---

load libraries
```{r}
library(tidyverse)
library(data.table)
library(gridExtra)
library(ggsci)
select <- dplyr::select
```

# Read in the stan model outputs - intermediate file saved later in code - can skip to ~line 180
```{r}
load("/home/kimberly.ledger/nearshore_eDNA/outputs/qm_field_samples/raw_20241122.Rdata")
raw <- Output

load("/home/kimberly.ledger/nearshore_eDNA/outputs/qm_field_samples/mock_20241122.Rdata")
mock1 <- Output 
```

```{r}
#########################################################
# Raw estimates from Reads
#########################################################
# summarize raw estimates from reads for each species.
raw.raw <- raw$env %>% 
  group_by(community,Cycles,tech_rep) %>%
  mutate(sum.ng = sum(start_conc_ng),
         true.prop = start_conc_ng / sum.ng) %>%
  ungroup() %>%
  group_by(Species,community,Cycles,true.prop) %>%
  group_by(Species,community,Cycles) %>%
  summarize(simple.Mean=mean(propReads),
            simple.N = length(tech_rep)) %>%
  replace_na(list(raw.Mean=0,raw.SD=0,raw.SE=0))

# extract predicted proportions from the posterior
COM <- data.frame(community = levels(raw$env$community %>% as.factor()))
COM$comm_idx <- 1:nrow(COM)
SP  <- raw$env %>% distinct(Species,sp_idx) %>% as.data.frame()

# These are the predicted intercepts for the posteriors
beta_posterior <- raw$stanMod_summary[["int_samp_small"]][, c(1,4:8)]
colnames(beta_posterior) <- paste0("raw.",substr(colnames(beta_posterior),1,nchar(colnames(beta_posterior))-1))
colnames(beta_posterior)[1] <- "raw.mean"
beta_posterior <- as.data.frame(beta_posterior)

raw.post <-expand.grid(comm_idx = COM$comm_idx,sp_idx =SP$sp_idx) %>% 
  arrange(comm_idx,sp_idx) %>% 
  left_join(.,COM) %>% 
  left_join(.,SP) %>% 
  bind_cols(.,beta_posterior)

# Combine the raw estimates and posterior estimates
raw.all <- full_join(raw.raw,raw.post)
```

```{r}
#########################################################
# Mock1
#########################################################
# summarize raw estimates from reads for each species.
mock1.raw <- mock1$env %>% group_by(community,Cycles,tech_rep) %>%
      mutate(sum.ng = sum(start_conc_ng),
             true.prop = start_conc_ng / sum.ng) %>%
      ungroup() %>%
      group_by(Species,community,Cycles,true.prop) %>%
  summarize(simple.Mean=mean(propReads),
            simple.N = length(tech_rep)) %>%
  replace_na(list(raw.Mean=0,raw.SD=0,raw.SE=0))

# extract predicted proportions from the posterior
COM <- data.frame(community = levels(mock1$env$community %>% as.factor()))
COM$comm_idx <- 1:nrow(COM)
SP  <- mock1$env %>% distinct(Species,sp_idx) %>% as.data.frame()

# These are the predicted intercepts for the posteriors
beta_posterior <- mock1$stanMod_summary[["int_samp_small"]][, c(1,4:8)]
colnames(beta_posterior) <- paste0("mock1.",substr(colnames(beta_posterior),1,nchar(colnames(beta_posterior))-1))
colnames(beta_posterior)[1] <- "mock1.mean"
beta_posterior <- as.data.frame(beta_posterior)

mock1.post <-expand.grid(comm_idx = COM$comm_idx,sp_idx =SP$sp_idx) %>% 
    arrange(comm_idx,sp_idx) %>% 
    left_join(.,COM) %>% 
    left_join(.,SP) %>% 
    bind_cols(.,beta_posterior)

# Combine the raw estimates and posterior estimates
mock1.all <- full_join(mock1.raw,mock1.post)
```

```{r}
# Combine mock results with raw reads.
result.dat <- left_join(mock1.all,raw.all)
```

plot estimates of amplification efficiency ( alpha)
```{r}
p_space_mock1 <- (exp(mock1$pars$alpha) / rowSums(exp(mock1$pars$alpha))) %>% as.data.frame()

colnames(p_space_mock1) <- mock1$Species

p_space_mock1_sum <- p_space_mock1 %>% 
                    pivot_longer( .,
                          cols = colnames(p_space_mock1),
                          names_to="species",values_to="val") %>%
                    group_by(species) %>%
                    summarize(Mean = mean(val),
                        SD=sd(val),
                        q.025 = quantile(val,probs=0.025),
                        q.05 = quantile(val,probs=0.05),
                        q.25 = quantile(val,probs=0.25),
                        q.75 = quantile(val,probs=0.75),
                        q.95 = quantile(val,probs=0.95),
                        q.975 = quantile(val,probs=0.975)) 

# get rid of reference species denotion
p_space_mock1_sum  <- p_space_mock1_sum  %>%
                      mutate(SP= ifelse(grepl("zRefSpecies_",species),
                                       substr(species,13,nchar(species)),
                                       as.character(species)))

plot1 <-  ggplot(p_space_mock1_sum) +
    geom_errorbarh(aes(xmin=q.25,xmax=q.75,y=SP),size=2,height=0) +
    geom_errorbarh(aes(xmin=q.025,xmax=q.975,y=SP),size=0.8,height=0) +
    geom_point(aes(x=Mean,y=SP),fill = "black", size=3,shape=21) +
    #scale_fill_manual(values=clr_alpha_sum$manual.col %>% as.character()) +
    scale_x_continuous("Amplification Efficiency (alpha)") +
    scale_y_discrete(NULL) +
    theme_bw() +
    theme(legend.position = "none",
          axis.text.y = element_text(size=10))

plot1

mean(p_space_mock1_sum$Mean)
```

clr
```{r}

clr_alpha_list_mock1 <- compositions::clr(p_space_mock1)

clr_alpha_mock1 <- clr_alpha_list_mock1 %>% as.data.frame() 

colnames(clr_alpha_mock1) <- mock1$Species

clr_alpha_sum <- clr_alpha_mock1 %>% 
                    pivot_longer( .,
                          cols = colnames(clr_alpha_mock1),
                          names_to="species",values_to="val") %>%
                    group_by(species) %>%
                    summarize(Mean = mean(val),
                        SD=sd(val),
                        q.025 = quantile(val,probs=0.025),
                        q.05 = quantile(val,probs=0.05),
                        q.25 = quantile(val,probs=0.25),
                        q.75 = quantile(val,probs=0.75),
                        q.95 = quantile(val,probs=0.95),
                        q.975 = quantile(val,probs=0.975)) 

# get rid of reference species denotion
clr_alpha_sum <- clr_alpha_sum %>%
                      mutate(SP= ifelse(grepl("zRefSpecies_",species),
                                       substr(species,13,nchar(species)),
                                       as.character(species)))

p_clr <-  ggplot(clr_alpha_sum) +
    geom_vline(xintercept=0,linetype="dashed") +
    geom_errorbarh(aes(xmin=q.25,xmax=q.75,y=SP),size=2,height=0) +
    geom_errorbarh(aes(xmin=q.025,xmax=q.975,y=SP),size=0.8,height=0) +
    geom_point(aes(x=Mean,y=SP,fill=SP),size=3,shape=21) +
    #scale_fill_manual(values=clr_alpha_sum$manual.col %>% as.character()) +
    scale_x_continuous("Amplification Efficiency (CLR)") +
    scale_y_discrete(NULL) +
    theme_bw() +
    theme(legend.position = "none",
          axis.text.y = element_text(size=10))

p_clr

mean(clr_alpha_sum$Mean)
```

alr
```{r}
alr_alpha_list_mock1 <- compositions::alr(p_space_mock1)

alr_alpha_mock1 <- alr_alpha_list_mock1 %>% as.data.frame() 

#colnames(alr_alpha_mock1) <- mock1$Species

alr_alpha_sum <- alr_alpha_mock1 %>% 
                    pivot_longer( .,
                          cols = colnames(alr_alpha_mock1),
                          names_to="species",values_to="val") %>%
                    group_by(species) %>%
                    summarize(Mean = mean(val),
                        SD=sd(val),
                        q.025 = quantile(val,probs=0.025),
                        q.05 = quantile(val,probs=0.05),
                        q.25 = quantile(val,probs=0.25),
                        q.75 = quantile(val,probs=0.75),
                        q.95 = quantile(val,probs=0.95),
                        q.975 = quantile(val,probs=0.975)) 

# get rid of reference species denotion
alr_alpha_sum <- alr_alpha_sum %>%
                      mutate(SP= ifelse(grepl("zRefSpecies_",species),
                                       substr(species,13,nchar(species)),
                                       as.character(species)))

p_alr <-  ggplot(alr_alpha_sum) +
    geom_vline(xintercept=0,linetype="dashed") +
    geom_errorbarh(aes(xmin=q.25,xmax=q.75,y=SP),size=2,height=0) +
    geom_errorbarh(aes(xmin=q.025,xmax=q.975,y=SP),size=0.8,height=0) +
    geom_point(aes(x=Mean,y=SP,fill=SP),size=3,shape=21) +
    #scale_fill_manual(values=clr_alpha_sum$manual.col %>% as.character()) +
    scale_x_continuous("Amplification Efficiency (ALR)") +
    scale_y_discrete(NULL) +
    theme_bw() +
    theme(legend.position = "none",
          axis.text.y = element_text(size=10))

p_alr

mean(clr_alpha_sum$Mean)
```


```{r}
ggsave(plot1, filename = "/home/kimberly.ledger/nearshore_eDNA/outputs/figures/alpha_all_mocks_combinedestimate.png", width = 7, height = 5, units = "in")
```


import sample metadata 
```{r}
metadata <- read.csv("/home/kimberly.ledger/nearshore_eDNA/data/nearshore_field_metadata.csv") %>%
  dplyr::rename(community = extraction_ID) %>%
  filter(sample_type == "sample") %>%
  select(community, alternative_ID, collection_year, bay_id, habitat)
```

join to qm
```{r}
result.join <- result.dat %>%
  left_join(metadata)

# get rid of reference species denotion
result.join <- result.join %>%
  mutate(SP= ifelse(grepl("zRefSpecies_",Species),
  substr(Species,13,nchar(Species)),
  as.character(Species)))

## rename yellowfin sole 
result.join <- result.join %>%
  mutate(Species = ifelse(Species == "Pleuronectidae 1", "Limanda aspera", Species), 
         SP = ifelse(SP == "Pleuronectidae 1", "Limanda aspera", SP))
```

save result.join
```{r}
#write.csv(result.join, "/home/kimberly.ledger/nearshore_eDNA/outputs/qm_field_samples/qm_results_20241125.csv")
```

read in qm results in needed 
```{r}
result.join <- read.csv("/home/kimberly.ledger/nearshore_eDNA/outputs/qm_field_samples/qm_results_20241125.csv") %>%
  select(!X)
```


plot raw estimates for REEF_A
```{r}
REEF_A_raw <- result.join %>%
  filter(bay_id == "REEF_A") %>%
  #filter(!is.na(simple.Mean)) %>%
  ggplot() + 
  geom_errorbar(aes(x=SP,ymin=raw.2.5,ymax=raw.97.5,color=SP),width=0) +
          geom_errorbar(aes(x=SP,ymin=raw.25,ymax=raw.75,color=SP),width=0, linewidth=2) +        
          geom_point(aes(x=SP,y=raw.mean,color=SP),shape =21,fill="black",size=2) +
          #scale_y_continuous(NULL,expand=c(0,NA),limits=c(0,NA),breaks=yBREAKS) +
          #scale_shape_manual(values=c(21,22,24)) +
          #scale_fill_manual(values=col.val,"Community") +
          #scale_color_manual(values=col.val,"Calibration") +
          #scale_x_continuous(NULL,breaks=xBREAKS$comm.idx,labels=xBREAKS$community,limits=c(NA,3.3)) +
  facet_wrap(~community) + 
          ylim(0, 1) +
          ylab("proportion of reads") +
          theme_classic() +
          theme(legend.position = "right",
                plot.margin = margin(0,0,0,0.85,"lines"),
                legend.key.size=unit(0.1,'lines'),
                legend.text=element_text(size=9),
                legend.title=element_text(size=10),
                axis.text.x = element_text(angle = 45, hjust = 1))
          
REEF_A_raw 
```

plot mock estimates for REEF_A
```{r}
REEF_A_mock1 <- result.join %>%
  filter(bay_id == "REEF_A") %>%
  #filter(!is.na(simple.Mean)) %>%
  ggplot() + 
  geom_errorbar(aes(x=SP,ymin=mock1.2.5,ymax=mock1.97.5,color=SP),width=0) +
          geom_errorbar(aes(x=SP,ymin=mock1.25,ymax=mock1.75,color=SP),width=0, linewidth=2) +        
          geom_point(aes(x=SP,y=mock1.mean,color=SP),shape =21,fill="black",size=2) +
          #scale_y_continuous(NULL,expand=c(0,NA),limits=c(0,NA),breaks=yBREAKS) +
          #scale_shape_manual(values=c(21,22,24)) +
          #scale_fill_manual(values=col.val,"Community") +
          #scale_color_manual(values=col.val,"Calibration") +
          #scale_x_continuous(NULL,breaks=xBREAKS$comm.idx,labels=xBREAKS$community,limits=c(NA,3.3)) +
  facet_wrap(~community) + 
          ylim(0, 1) +
          ylab("proportion of reads") +
          theme_classic() +
          theme(legend.position = "right",
                plot.margin = margin(0,0,0,0.85,"lines"),
                legend.key.size=unit(0.1,'lines'),
                legend.text=element_text(size=9),
                legend.title=element_text(size=10),
                axis.text.x = element_text(angle = 45, hjust = 1))
          
REEF_A_mock1 
```

calculate dissimilarity across biological replicates 

how many biological replicates represent each bay? 
```{r}
result.join  %>%
  group_by(collection_year, bay_id) %>%
  summarise(nrep = n_distinct(community))
```

great. all have two or three biological replicates 

```{r}
result.join <- result.join %>%
  unite(year_bay, collection_year, bay_id, sep = "_", remove = F) %>%
  unite(bay_comm,  year_bay, community, sep = "-", remove = F)
```

```{r}
library(vegan)

tibble_to_matrix <- function (tb) {
  
  tb %>%
  #result.join %>%
    group_by(bay_comm, SP) %>% 
    summarise(mock1.mean = sum(mock1.mean)) %>% 
    spread ( key = "SP", value = "mock1.mean", fill = 0) %>%
    ungroup() -> matrix_1
    samples <- pull (matrix_1, bay_comm)
    matrix_1[,-1] -> matrix_1
    data.matrix(matrix_1) -> matrix_1
    dimnames(matrix_1)[[1]] <- samples
    vegdist(matrix_1) -> matrix_1
}
```

```{r}
all.distances.full <- tibble_to_matrix(result.join)

# Do all samples have a name?
summary(is.na(names(all.distances.full)))
```

make the pairwise distances a long table
```{r}
as_tibble(subset(melt(as.matrix(all.distances.full)))) -> all.distances.melted

# Any major screw ups
summary(is.na(all.distances.melted$value))

# Now, create a three variables for all distances, they could be PCR replicates, BIOL replicates, or from the same site

all.distances.melted %>%
  separate (Var1, into = c("Site1", "community"), sep = "-", remove = FALSE) %>%
  separate (Var2, into = c("Site2", "community"), sep = "-", remove = FALSE) %>%
  mutate (Distance.type = case_when(Site1 == Site2 ~ "Same Site",
                                      TRUE ~ "Different Site")) %>%
  #dplyr::select(Sample1 = Var1, Sample2 = Var2 , value , Distance.type) %>%
  filter (Var1 != Var2) -> all.distances.to.plot

# Checking all went well

sapply(all.distances.to.plot, function(x) summary(is.na(x)))
```

```{r}
#all.distances.to.plot$Distance.type <- all.distances.to.plot$Distance.type  %>% fct_relevel("PCR Replicates", "Same Site")

ggplot (all.distances.to.plot) +
  geom_histogram (aes (fill = Distance.type, x = value, after_stat(ndensity)), position = "dodge",  alpha = 0.9, bins = 50) +
  facet_wrap( ~ Distance.type) +
  labs (x = "Pairwise dissimilarity", y = "density" ,
        Distance.type = "Distance",
        title = "") +
    guides (fill = "none")
```

what are the sites with the pairwise dissimilarity above 0.75? 
```{r}
all.distances.to.plot %>%
  filter(Distance.type == "Same Site") %>%
  filter(value > 0.75)
```


load biomass data to compare w/ qm read proportions 
```{r}
biomass <- read.csv("/home/kimberly.ledger/nearshore_eDNA/data/seine_biomass_at_edna_sites.csv") %>%
  select(!X)

unique(biomass$species_scientific)
```

filter for just the mock community species - change spp names to match metabarcoding
```{r}
mock_spp <- c("Ammodytes personatus", "Blepsias cirrhosus", "Clupea pallasii", "Cymatogaster aggregata", "Gadus macrocephalus", "Gadus chalcogrammus", "Gasterosteus aculeatus", "Hexagrammos stelleri", "Hexagrammos octogrammus", "Hexagrammos lagocephalus", "Lepidopsetta spp.", "Leptocottus armatus", "Limanda aspera", "Lumpenus sagitta", "Oncorhynchus keta", "Oncorhynchus kisutch", "Oncorhynchus gorbuscha", "Ophiodon elongatus", "Pholis laeta", "Sebastes auriculatus", "Sebastes melanops", "Sebastes sp", "Sebastes caurinus", "Sebastes maliger", "Syngnathus leptorhynchus")

biomass_mock_spp <- biomass %>%
  filter(species_scientific %in% mock_spp) %>%
  mutate(SP = species_scientific) %>%
  mutate(SP = ifelse(SP %in% c("Gadus macrocephalus", "Gadus chalcogrammus"), "Gadus", SP),
         SP = ifelse(SP %in% c("Hexagrammos stelleri", "Hexagrammos octogrammus", "Hexagrammos lagocephalus"), "Hexagrammos", SP),
         SP = ifelse(SP == "Lepidopsetta spp.", "Lepidopsetta", SP),
         #SP = ifelse(SP == "Limanda aspera", "Pleuronectidae 1", SP),  ## i am changed the ID in the mb data
         SP = ifelse(SP %in% c("Sebastes caurinus", "Sebastes maliger"), "Sebastes 1", SP),
         SP = ifelse(SP == "Sebastes auriculatus", "Sebastes 2", SP),
         SP = ifelse(SP %in% c("Sebastes melanops", "Sebastes sp"), "Sebastes 3", SP))

unique(biomass_mock_spp$SP)
unique(biomass_mock_spp$bay_id)
```

calculate biomass proportions 
```{r}
biomass_df <- biomass_mock_spp %>%
  group_by(bay_id, year, habitat, SP) %>%
  summarize(biomass_g = sum(mass_g)) %>%
  ungroup() %>%
  group_by(bay_id, year, habitat) %>%
  mutate(sum.biomass = sum(biomass_g),
             biomass.prop = biomass_g / sum.biomass)
```

make barplots of biomass proportions 
```{r}
plot_biomass_2021 <- biomass_df %>%
  filter(year == 2021) %>%
  ggplot(aes(x=bay_id, y=biomass.prop, fill=SP)) +
  geom_bar(stat = "identity") + 
  facet_wrap(~ habitat, scales = "free_x") +
  theme_bw() +
  labs(
    y = "biomass (%)",
    x = "",
    title = "") + 
  theme(
    axis.text.x = element_text(angle = 90, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )

plot_biomass_2021
```

```{r}
plot_biomass_2022 <- biomass_df %>%
  filter(year == 2022) %>%
  ggplot(aes(x=bay_id, y=biomass.prop, fill=SP)) +
  geom_bar(stat = "identity") + 
  facet_wrap(~ habitat, scales = "free_x") +
  theme_bw() +
  labs(
    y = "biomass (%)",
    x = "",
    title = "") + 
  theme(
    axis.text.x = element_text(angle = 90, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )

plot_biomass_2022
```


combine data frames so that can make side-by-side biomass and qm read proportion plots for sites 
- it looks like a few habitat types were incorrectly entered into eDNA db... i'll match them to biomass habitat categories for now but should double check they are correct 
```{r}
df_mb <- result.join %>%
  ungroup() %>%
  mutate(habitat = ifelse(bay_id == "BLAQ_A", "eelgrass", habitat),
         habitat = ifelse(bay_id == "ROSA_A", "mixed", habitat),
         habitat = ifelse(bay_id == "SFEI_A", "eelgrass", habitat),
         habitat = ifelse(bay_id == "TRIS_A", "mixed", habitat)) %>%
  select(collection_year, bay_id, habitat, community, SP, mock1.mean, raw.mean) %>%
  rename(year = collection_year) %>%
  pivot_longer(cols=c(6:7), names_to = "method", values_to = "prop")

db_biomass <- biomass_df %>%
  select(!biomass_g) %>%
  select(!sum.biomass) %>%
  mutate(method = "seine") %>%
  mutate(community = NA) %>%
  rename(prop = biomass.prop)
  
df<- df_mb %>%
  bind_rows(db_biomass)

df %>% select(year, bay_id) %>% unique() %>% arrange(bay_id)
```

note: CLAM_A and KLWA_A were sampled in both 2021 and 2022

plot bay of your choice to compare eDNA mb and seine biomass proportions 
```{r}
df_filter <- df %>%
  filter(method != "raw.mean") %>%
  filter(bay_id == "SYLB_A")
  
# Extract the unique values after filtering
year <- unique(df_filter$year)
bay <- unique(df_filter$bay_id)
habitat <- unique(df_filter$habitat)
  
df_filter %>%
  ggplot(aes(x=community, y=prop, fill=SP)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  facet_grid(~method, scales = 'free') +
  labs(
    y = "proportion",
    x = "method") +
  ggtitle(paste("Year:", year, "; Bay:", bay, "; Habitat:", habitat)) + 
  theme(
    axis.text.x = element_text(angle = 90, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )
```


```{r}
df_filter <- df %>%
  filter(method != "raw.mean") %>%
  filter(bay_id == "SHIN_A")
  
# Extract the unique values after filtering
year <- unique(df_filter$year)
bay <- unique(df_filter$bay_id)
habitat <- unique(df_filter$habitat)
  
df_filter %>%
  ggplot(aes(x=community, y=prop, fill=SP)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  facet_grid(~method, scales = 'free') +
  labs(
    y = "proportion",
    x = "method") +
  ggtitle(paste("Year:", year, "; Bay:", bay, "; Habitat:", habitat)) + 
  theme(
    axis.text.x = element_text(angle = 90, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )
```

```{r}
df_filter <- df %>%
  filter(method != "raw.mean") %>%
  filter(bay_id == "REEF_A")
  
# Extract the unique values after filtering
year <- unique(df_filter$year)
bay <- unique(df_filter$bay_id)
habitat <- unique(df_filter$habitat)
  
df_filter %>%
  ggplot(aes(x=community, y=prop, fill=SP)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  facet_grid(~method, scales = 'free') +
  labs(
    y = "proportion",
    x = "method") +
  ggtitle(paste("Year:", year, "; Bay:", bay, "; Habitat:", habitat)) + 
  theme(
    axis.text.x = element_text(angle = 90, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )
```


plot everything... mb biologial reps are seperate here 
```{r}
df %>%
  filter(method != "raw.mean") %>%
  mutate(community = ifelse(is.na(community), paste(year, bay_id), community)) %>%
  ggplot(aes(x=community, y=prop, fill=SP)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  facet_grid(~method, scales = 'free') +
  labs(
    y = "proportion",
    x = "method") +
  theme(
    axis.text.x = element_text(angle = 90, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )
```

too much to look at all at once.

calculate mean of qm read proportions per bay - DO WE WANT TO DO THIS? 

also going to use the RAW proportions - ie. not qm adjusted 
```{r}
df_bay_means <- df_mb %>%
  #filter(method == "mock1.mean") %>%
  filter(method == "raw.mean") %>%
  group_by(year, bay_id, habitat, SP) %>%
  summarize(mean_prop = mean(prop),
            SD=sd(prop),
                q.025 = quantile(prop,probs=0.025),
                Median = quantile(prop,probs=0.5),
                q.975 = quantile(prop,probs=0.975)) %>%
  mutate(method = "raw_mean_of_biological_reps")
```

do means then sum to 1? 
```{r}
df_bay_means %>%
  group_by(year, bay_id, habitat) %>%
  summarise(total = sum(mean_prop))
```


plot 
```{r}
df_bay_means %>%
  unite(year_bay, year, bay_id, sep = "_", remove = F) %>%
  ggplot(aes(x=year_bay, y=mean_prop, fill=SP)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  facet_grid(~method, scales = 'free') +
  labs(
    y = "proportion",
    x = "method") +
  theme(
    axis.text.x = element_text(angle = 90, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )
```

join this with the biomass data 
```{r}
df2 <- df_bay_means %>%
  rename(prop = mean_prop) %>%
  bind_rows(db_biomass) %>%
  select(!community) %>%
  unite(year_bay, year, bay_id, sep = "_", remove = F)
```


plots 
```{r}
fig_2021 <- df2 %>%
  filter(year == 2021) %>%
  ggplot(aes(x=year_bay, y=prop, fill=SP)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  facet_grid(~method, scales = 'free') +
  labs(
    y = "proportion",
    x = "method") +
  theme(
    axis.text.x = element_text(angle = 90, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )

fig_2021

#ggsave(plot = fig_2021, "outputs/samples_w_mocks_updated/figures/barplots_2021.png", dpi = 300, width = 10, height = 6)
```

```{r}
fig_2022 <- df2 %>%
  filter(year == 2022) %>%
  ggplot(aes(x=year_bay, y=prop, fill=SP)) +
  geom_bar(stat = "identity") + 
  theme_bw() +
  facet_grid(~method, scales = 'free') +
  labs(
    y = "proportion",
    x = "method") +
  theme(
    axis.text.x = element_text(angle = 90, hjust = 0.95),
    legend.text = element_text(size = 8),
    legend.key.size = unit(0.3, "cm"),
    legend.position = "right",
    legend.title = element_blank()
  )
fig_2022

#ggsave(plot = fig_2022, "outputs/samples_w_mocks_updated/figures/barplots_2022.png", dpi = 300, width = 10, height = 6)
```


scatterplots? 
```{r}
df3 <- df2 %>%
  pivot_wider(names_from = "method", values_from = c("prop":"q.975")) %>%
  select(!SD_seine) %>%
  select(!q.025_seine) %>%
  select(!Median_seine) %>%
  select(!q.975_seine) %>%
  #mutate(qm = ifelse(qm_mean_of_biological_reps < 0.01, 0, qm_mean_of_biological_reps)) %>%
  mutate(prop_raw_mean_of_biological_reps = ifelse(prop_raw_mean_of_biological_reps < 0.01, 0, prop_raw_mean_of_biological_reps)) %>%
    mutate(SD_raw_mean_of_biological_reps = ifelse(prop_raw_mean_of_biological_reps < 0.01, 0, SD_raw_mean_of_biological_reps)) %>%
    mutate(q.025_raw_mean_of_biological_reps = ifelse(prop_raw_mean_of_biological_reps < 0.01, 0, q.025_raw_mean_of_biological_reps)) %>%
    mutate(Median_raw_mean_of_biological_reps = ifelse(prop_raw_mean_of_biological_reps < 0.01, 0, Median_raw_mean_of_biological_reps)) %>%
    mutate(q.975_raw_mean_of_biological_reps = ifelse(prop_raw_mean_of_biological_reps < 0.01, 0, q.975_raw_mean_of_biological_reps)) %>%
  mutate(prop_seine = ifelse(is.na(prop_seine), 0, prop_seine))
```


```{r}
fig <- df3 %>%
  ggplot() +
  geom_errorbar(aes(x=prop_seine, ymin=q.025_raw_mean_of_biological_reps, ymax=q.975_raw_mean_of_biological_reps, color=SP),width=0, size = .5) +
  geom_point(aes(x= prop_seine, y = prop_raw_mean_of_biological_reps, color = SP)) + 
  theme_bw() + 
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  +
  facet_wrap(~SP, scales = "free") #+
  #xlim(0,1) +
  #ylim(0,1)

fig
  

ggsave(plot = fig, "/home/kimberly.ledger/nearshore_eDNA/outputs/figures/biomass_v_qm_zoom_95IQR.png", dpi = 300, width = 10, height = 6)
```

```{r}
fig2 <- df3 %>%
  ggplot() +
  geom_errorbar(aes(x=prop_seine, ymin=q.025_raw_mean_of_biological_reps, ymax=q.975_raw_mean_of_biological_reps, color=habitat),width=0, size = .5) +
  geom_point(aes(x= prop_seine, y = prop_raw_mean_of_biological_reps, color = habitat)) + 
  theme_bw() + 
  geom_abline(intercept = 0, slope = 1, linetype = "dashed", color = "black")  +
  facet_wrap(~SP, scales = "free") +
  scale_x_continuous(limits = c(0, NA)) +  # Limit x-axis to non-negative values
  scale_y_continuous(limits = c(0, NA)) #+  # Limit y-axis to non-negative values
  #xlim(0,1) +
  #ylim(0,1)

fig2
  
#ggsave(plot = fig2, "/home/kimberly.ledger/nearshore_eDNA/outputs/figures/biomass_v_qm_by_habitat.png", dpi = 300, width = 10, height = 6)
```
