==> code/7_quantmetabarcoding_fieldsamples.Rmd <==
---
title: "Quantitative metabarcoding of AK nearshore field samples"
author:"Kimberly Ledger"
date: "2024-11-13"
output: github_document
---

this code is using the gblock mock communities to correct for amplification biases in the nearshore eDNA field samples

libraries
```{r}
library(tidyverse)
library(compositions)
  select <- dplyr::select  #eliminate conflict w MASS::select vs. dplyr::select
  rename <- dplyr::rename
library(rstan)
```

load mock community data
```{r}
mockcom_data <- read.csv("/home/kimberly.ledger/nearshore_eDNA/outputs/samples_w_mocks_updated/ns_mock_20241119.csv") %>%
  mutate(tech_rep = as.factor(tech_rep),   ## i changed this from as.numeric to as.factor because of my rep format
         Cycles = as.numeric(Cycles)) %>%
  filter(start_conc_ng > 0) %>%   ## remove any reads belonging to species NOT in the mock community - think about whether this is the best approach or if i should do some "decontamination" steps to all the mock community reads 
  filter(community != "COM4") %>%  ## removing because i'm not including communities with chinook/coho
  mutate(nReads = as.integer(nReads))
```

unique taxa?
```{r}
mockcom_data %>%
  select(taxon) %>%
  unique() %>%
  arrange(taxon)
```

group species by mock community and calculate total concentration
```{r}
sp.by.comm <- mockcom_data %>% filter(Cycles==35) %>% 
  group_by(species=taxon,community) %>%
  summarize(conc = sum(start_conc_ng)) %>%
  mutate(conc2 = ifelse(conc>0,1,0)) %>%
  arrange(species,community) %>% group_by(community) %>%
  mutate(tot_conc = sum(conc),prop=conc/tot_conc)
```

#REQUIRED FORMAT
#environmental (i.e., unknown) samples with the following column names:
  #community  -- the unique biological sample from which a set of sequences derived
  #Species   -- the biological species to which reads are assigned
  #nReads -- the number of reads assigned to that species in that community
  #tech_rep -- index of technical replicate, within community (e.g., if each community is sequenced 3 times, tech_rep will vary between 1 and 3)
  #Cycles -- number of PCR cycles the reaction underwent
  
  
load field samples 
```{r}
decontaminated_table <- read.csv("/home/kimberly.ledger/nearshore_eDNA/outputs/samples_w_mocks_updated/decontaminated_asv_table.csv") %>%
  select(!X) %>%
  filter(sample_type == "sample") %>%
  filter(reads > 0)
```

assign asv's to taxons 
```{r}
taxonomy <- read.csv("/home/kimberly.ledger/nearshore_eDNA/outputs/samples_w_mocks_updated/taxonomy_20241122_collapsed.csv") %>%
  select(!X) %>%
  select(qseqid, taxon) %>%
  unique() %>%
  rename(ASV = qseqid) 
```

join to decontaminated ASV table
```{r}
decon_table_w_tax <-  decontaminated_table %>%
  left_join(taxonomy, by = "ASV")
```

remove ASV IDs and group by taxon 
```{r}
taxon_table <- decon_table_w_tax %>%
  group_by(Sample_ID, taxon) %>%
  summarise(nReads = sum(reads)) %>%
  mutate(nReads = as.integer(nReads)) %>%
  separate(Sample_ID, into = c("community", "tech_rep")) %>%
  rename(Species = taxon) %>%
  mutate(Cycles = 35) %>%
  mutate(tech_rep = ifelse(tech_rep == "A", "1", tech_rep)) %>%
  mutate(tech_rep = ifelse(tech_rep == "B", "2", tech_rep)) %>%
  mutate(tech_rep = ifelse(tech_rep == "C", "3", tech_rep)) %>%
  mutate(tech_rep = as.integer(tech_rep))

# Check for samples missing pcr_replicate 1
missing_replicate1 <- taxon_table %>%
  group_by(community) %>%
  filter(!1 %in% tech_rep) %>%
  select(community, tech_rep) %>%
  unique() %>%
  filter(tech_rep == 3)  ## will just change the 3's to be 1's 

# now mutate rows
taxon_table_temp <- taxon_table %>%
  mutate(tech_rep = ifelse(paste(community, tech_rep) %in% paste(missing_replicate1$community, missing_replicate1$tech_rep), 1L, tech_rep))

# Check for samples missing pcr_replicate 1
missing_replicate1_again <- taxon_table_temp%>%
  group_by(community) %>%
  filter(!1 %in% tech_rep)

#okay good, now renaming df
taxon_table <- taxon_table_temp
```

  
```{r}
taxon_table_formatted <- taxon_table %>%
  mutate(start_conc_ng = NA) 
  
env <- taxon_table_formatted %>%  
  group_by(community, tech_rep) %>% 
  mutate(propReads = nReads/sum(nReads), #calculate proportion of reads
          totReads = sum(nReads)) %>%  #calculate total reads for community
  group_by(Species) %>% 
  mutate(totalSpeciesReads = sum(nReads)) %>%  
  add_tally(nReads > 0, name = "totalOccurrences") %>% 
  filter(totalSpeciesReads > 0)
         
#assign most common species to be the reference species
mostCommon <- env %>% 
    group_by(Species) %>% 
    tally(nReads > 0) %>%
    arrange(desc(n)) %>% 
    head(1) %>% 
    pull(Species)
env$Species[env$Species == mostCommon] <- paste0("zRefSpecies_", mostCommon) #make the most common species the reference species
env <- env %>% 
    arrange(Species, community)
```

#REQUIRED FORMAT
mock community (i.e., known) samples with the following column names:
  #community  -- the unique biological sample from which a set of sequences derived
  #Species   -- the biological species to which reads are assigned
  #nReads -- the number of reads assigned to that species in that community
  #tech_rep -- index of technical replicate, within community (e.g., if each community is sequenced 3 times, tech_rep will vary between 1 and 3)
  #Cycles -- number of PCR cycles the reaction underwent
  #start_conc_ng -- starting concentration of each species in each community, in nanograms

```{r}
#set up mock community samples
mc <- mockcom_data %>%
  mutate(tech_rep = ifelse(tech_rep == "A", "1", tech_rep)) %>%
  mutate(tech_rep = ifelse(tech_rep == "B", "2", tech_rep)) %>%
  mutate(tech_rep = ifelse(tech_rep == "C", "3", tech_rep)) %>%
  mutate(tech_rep = as.integer(tech_rep)) %>%
  rename("Species" = "taxon") #mock comm samples
  
mc$Species[mc$Species == mostCommon] <- paste0("zRefSpecies_", mostCommon)   #make the most common species the reference species
  
# Filter so that you only keep species in the environment samples that are in the mock community.
# It is ok to include species that are only in the mock community.
env <- env %>%
    filter(Species %in% mc$Species)%>% #limit to species occurring in mock community dataset
    arrange(Species, community)  
  
#double check
sum(!mc$Species %in% unique(env$Species)) # This can be non-zero
sum(!env$Species %in% unique(mc$Species)) # this had better be zero.
```

```{r}
# Make a single species list:
  sp.list   <- data.frame(Species = sort(unique(mc$Species)) ) %>% mutate(sp_idx =1:length(Species))
  N_species <- nrow(sp.list)
    
  comm.mock.list <- mc %>% group_by(community, tech_rep,Cycles) %>% summarize(n=length(tech_rep)) %>%
                        ungroup() %>% mutate(id=1:length(n))
  comm.env.list   <- env %>% group_by(community, tech_rep,Cycles) %>% summarize(n=length(tech_rep)) %>%
                        ungroup() %>% mutate(id=1:length(n))
    
    #make a list of species that are in mock community but not environment, 
    # expand grid to make it so the the environmental samples get padded with all the
    # missing species for all of the communities and all tech replicates.
    
    sp.comm.mc  <- expand_grid(Species = sp.list$Species, id = comm.mock.list$id) %>% 
                          left_join(.,sp.list %>% select(Species,sp_idx)) %>%
                          left_join(.,comm.mock.list %>% select(community,tech_rep,Cycles,id) ) %>% select(-id)
    sp.comm.env <- expand_grid(Species = sp.list$Species, id = comm.env.list$id) %>% 
                          left_join(.,sp.list %>% select(Species,sp_idx)) %>%
                          left_join(.,comm.env.list %>% select(community,tech_rep,Cycles,id) ) %>% select(-id)

    

    #convert to matrices
    # merge in species and indices first to make pivoting more efficient.
    
    mc  <- left_join(sp.comm.mc,mc) %>%   
              mutate(nReads = ifelse(is.na(nReads),0,nReads),
                  start_conc_ng = ifelse(is.na(start_conc_ng),0,start_conc_ng)) 
    env <- left_join(sp.comm.env,env) %>%
              mutate(nReads = ifelse(is.na(nReads),0,nReads),
                  start_conc_ng = ifelse(is.na(start_conc_ng),0,start_conc_ng))
    
    sample_data <- env %>% 
      ungroup() %>% 
      dplyr::select(community, sp_idx, nReads, tech_rep, Cycles) %>% 
      arrange(sp_idx) %>% 
      pivot_wider(names_from = "sp_idx", values_from = "nReads", values_fill = 0) 
    
    sample_data_small <- sample_data %>% filter(tech_rep== "1")
    
    mock_data <- mc %>% 
      ungroup() %>% 
      dplyr::select(community, sp_idx, nReads,tech_rep, Cycles) %>% 
      arrange(sp_idx) %>% 
      pivot_wider(names_from = "sp_idx", values_from = "nReads", values_fill = 0)

    mock_data_small <- mock_data %>% filter(tech_rep== "1")
```


```{r}
#proportions
p_mock <- mc %>% 
  select(community, tech_rep, sp_idx, start_conc_ng, Cycles) %>% 
  arrange(sp_idx) %>% 
  group_by(community, tech_rep, Cycles) %>% 
  mutate(prop_conc = start_conc_ng/sum(start_conc_ng)) %>% 
  select(-start_conc_ng) %>% #, -Species) %>% 
  pivot_wider(names_from = "sp_idx", values_from = "prop_conc", values_fill = 0) %>% 
  ungroup() %>% 
  arrange(community)
  #select(-community)

p_mock_small <- mc %>%
  filter(tech_rep == "1") %>%
  select(community, sp_idx, start_conc_ng, Cycles) %>%
  arrange(sp_idx) %>%
  group_by(community) %>%
  mutate(prop_conc = start_conc_ng/sum(start_conc_ng)) %>%
  select(-start_conc_ng) %>%  # -Species) %>%
  pivot_wider(names_from = "sp_idx", values_from = "prop_conc", values_fill = 0) %>%
  ungroup() %>%
  arrange(community)
  #select(-community, -Cycles)

  #calculate additive log ratios 
  alr_mock_true_prop <- p_mock[,4:(ncol(p_mock)-1)]*0
  #i=1
  for(i in 1:nrow(p_mock)){
    alr_mock_true_prop[i,] <- compositions::alr(p_mock[i,4:(ncol(p_mock))] + 1e-10)
  }
  alr_mock_true_prop[,N_species] <- 0 #adding explicit reference species column
  
  alr_mock_true_prop_small <- p_mock_small[,3:(ncol(p_mock_small)-1)]*0
  for(i in 1:nrow(p_mock_small)){
    alr_mock_true_prop_small[i,] <- compositions::alr(p_mock_small[i,3:(ncol(p_mock_small))] + 1e-10)
  }
  alr_mock_true_prop_small[,N_species] <- 0
```


#DESIGN MATRICES
  # mock communities first
  # species compositions (betas)
  # use mock_data  
```{r}
  N_pcr_mock <- mock_data$Cycles
  
  if(length(unique(mock_data$community))==1){
    formula_b <- Cycles ~ 1  # what is on the left side of the equation doesn't matter.
  } else {
    formula_b <- Cycles ~ community # what is on the left side of the equation doesn't matter.
  }
  model_frame <- model.frame(formula_b, mock_data)
  model_matrix_b_mock <- model.matrix(formula_b, model_frame)
  
  #formula_b <- obs_sp_1 ~ community
  model_frame <- model.frame(formula_b, mock_data_small)
  model_matrix_b_mock_small <- model.matrix(formula_b, model_frame)
  
  # efficiencies (alphas)
  formula_a <- community ~ Cycles -1
  model_frame <- model.frame(formula_a, mock_data)
  model_vector_a_mock <- model.matrix(formula_a, model_frame) %>% as.numeric()
  model_frame <- model.frame(formula_a, mock_data_small)
  model_vector_a_mock_small <- model.matrix(formula_a, model_frame) %>% as.numeric()
  
  N_obs_mock_small <- nrow(model_matrix_b_mock_small)
  N_obs_mock       <- nrow(mock_data)
  N_b_mock_col     <- ncol(model_matrix_b_mock)  
  
  # unknown communities second
  # species compositions (betas)

  # use sample_data
  
  N_pcr_samp <- sample_data$Cycles
  
  if(length(unique(sample_data$community))==1){
    formula_b <- Cycles ~ 1  
  } else {
    formula_b <- Cycles ~ community
  }
  model_frame <- model.frame(formula_b, sample_data)
  model_matrix_b_samp <- model.matrix(formula_b, model_frame)
  
  #formula_b <- obs_sp_1 ~ community 
  #p_samp_all$site <- as.factor(p_samp_all$site)
  model_frame <- model.frame(formula_b, sample_data_small)
  model_matrix_b_samp_small <- model.matrix(formula_b, model_frame)
  
  # efficiencies (alpha)
  formula_a <- community ~ Cycles -1
  model_frame <- model.frame(formula_a, sample_data)
  model_vector_a_samp <- model.matrix(formula_a, model_frame) %>% as.numeric()
  model_frame <- model.frame(formula_a, sample_data_small)
  model_vector_a_samp_small <- model.matrix(formula_a, model_frame) %>% as.numeric()
  
  #counters 
  N_obs_samp_small <- nrow(model_matrix_b_samp_small)
  N_obs_samp <- nrow(sample_data)
  N_b_samp_col <- ncol(model_matrix_b_samp)               
  
```

set up for stan model
```{r}
stan_data <- list(
  N_species = N_species,   # Number of species in data
  N_obs_samp = N_obs_samp, # Number of observed samples 
  N_obs_mock = N_obs_mock, # Number of observed mock samples
  N_obs_samp_small = N_obs_samp_small, # Number of observed samples 
  N_obs_mock_small = N_obs_mock_small, # Number of observed mock samples
  
  # Observed data of community matrices
  sample_data = sample_data %>% select(-community,-Cycles,-tech_rep),
  mock_data   = mock_data  %>% select(-community,-Cycles,-tech_rep),
  
  # True proportions for mock community
  #mock_true_prop = p_mock_all %>% dplyr::select(contains("sp")),
  alr_mock_true_prop = alr_mock_true_prop,
  alr_mock_true_prop_small = alr_mock_true_prop_small,
  
  # vectors of PCR numbers
  N_pcr_samp = N_pcr_samp,
  N_pcr_mock = N_pcr_mock,
  
  # Design matrices: field samples
  N_b_samp_col = N_b_samp_col,
  model_matrix_b_samp = model_matrix_b_samp,
  model_matrix_b_samp_small = model_matrix_b_samp_small,
  model_vector_a_samp = model_vector_a_samp,
  model_vector_a_samp_small = as.array(model_vector_a_samp_small),
  
  # Design matrices: mock community samples
  N_b_mock_col = N_b_mock_col,
  model_matrix_b_mock = model_matrix_b_mock,
  model_matrix_b_mock_small = model_matrix_b_mock_small,
  model_vector_a_mock = model_vector_a_mock,
  model_vector_a_mock_small = model_vector_a_mock_small,
  
  # Priors
  alpha_prior = c(0,0.1),  # normal prior
  beta_prior = c(0,10),    # normal prior
  tau_prior = c(1.5,1.5)   # gamma prior
)


stan_pars <- c(
  #"alpha",
  "beta",
  "eta_samp",
  #"eta_mock",
  "tau",
  "mu_samp",
  #"mu_mock",
  "int_samp_small"
)
```

```{r}
#########################################
#########################################
#Bayesian Estimation
N_CHAIN = 3
Warm = 1000
Iter = 1500
Treedepth = 15
Adapt_delta = 0.70

rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())

setwd("~/nwfsc-cb-Toward_Quantitative_Metabarcoding-fba0076/code/Core_Stan_models")
## my note: this would not run without having the wd set to the Core_Stan_models folder... 

stanMod = stan(file = "quant_metabar_no_mock_no_alpha.stan" ,data = stan_data,
               verbose = FALSE, chains = N_CHAIN, thin = 1,
               warmup = Warm, iter = Warm + Iter,
               control = list(adapt_init_buffer = 175,
                              max_treedepth=Treedepth,
                              stepsize=0.01,
                              adapt_delta=Adapt_delta,
                              metric="diag_e"),
               pars = stan_pars,
               refresh = 10,
               boost_lib = NULL,
               #init = stan_init_f2(n.chain=N_CHAIN,N_species=N_species),
               sample_file = paste0("./tmpE.csv")
)

pars <- rstan::extract(stanMod, permuted = TRUE)
samp_params <- get_sampler_params(stanMod)

stanMod_summary <- list()
stanMod_summary[["tau"]] <- summary(stanMod,pars=c("tau"))$summary
stanMod_summary[["beta"]] <- summary(stanMod,pars="beta")$summary
stanMod_summary[["eta_samp_raw"]] <- summary(stanMod,pars="eta_samp")$summary
stanMod_summary[["mu_samp"]] <- summary(stanMod,pars="mu_samp")$summary
stanMod_summary[["int_samp_small"]] <- summary(stanMod,pars="int_samp_small")$summary


Output <- list(
  mockcom_data = mockcom_data, # raw data from all observations, all communities.
  env = env,  #environmental data
  mc = mc, #mock data
  Species = unique(mc$Species),
  
  # Realizations of simulations (input data)
  p_true = p_mock,
  p_samp_all = sample_data,
  p_mock_all = mock_data,
  
  # stan input objects
  stan_data = stan_data,
  Warm=Warm,
  Iter=Iter,
  
  # Fitted Objects
  stanMod = stanMod, # Full stan Model fitted object
  pars = pars, # MCMC output
  samp_params=samp_params, # Sampler information
  stanMod_summary = stanMod_summary # posterior summaries.
)
```

```{r}
save(Output,file=paste0("/home/kimberly.ledger/nearshore_eDNA/outputs/samples_w_mocks_updated/field_quant/raw_20241122",".Rdata"))
```


## now run stan models that adjust for amplification efficiency 

  #DESIGN MATRICES
  # mock communities first
  # species compositions (betas)
  # use mock_data  
```{r}
  N_pcr_mock <- mock_data$Cycles
  
  if(length(unique(mock_data$community))==1){
    formula_b <- Cycles ~ 1  # what is on the left side of the equation doesn't matter.
  } else {
    formula_b <- Cycles ~ community # what is on the left side of the equation doesn't matter.
  }
  model_frame <- model.frame(formula_b, mock_data)
  model_matrix_b_mock <- model.matrix(formula_b, model_frame)
  
  #formula_b <- obs_sp_1 ~ community 
  model_frame <- model.frame(formula_b, mock_data_small)
  model_matrix_b_mock_small <- model.matrix(formula_b, model_frame)
  
  # efficiencies (alphas)
  formula_a <- community ~ Cycles -1
  model_frame <- model.frame(formula_a, mock_data)
  model_vector_a_mock <- model.matrix(formula_a, model_frame) %>% as.numeric()
  model_frame <- model.frame(formula_a, mock_data_small)
  model_vector_a_mock_small <- model.matrix(formula_a, model_frame) %>% as.numeric()
  
  N_obs_mock_small <- nrow(model_matrix_b_mock_small)
  N_obs_mock       <- nrow(mock_data)
  N_b_mock_col     <- ncol(model_matrix_b_mock)  
  
  # unknown communities second
  # species compositions (betas)

  # use sample_data
  
  N_pcr_samp <- sample_data$Cycles
  
  if(length(unique(sample_data$community))==1){
    formula_b <- Cycles ~ 1  
  } else {
    formula_b <- Cycles ~ community
  }
  model_frame <- model.frame(formula_b, sample_data)
  model_matrix_b_samp <- model.matrix(formula_b, model_frame)
  
  #formula_b <- obs_sp_1 ~ community 
  #p_samp_all$site <- as.factor(p_samp_all$site)
  model_frame <- model.frame(formula_b, sample_data_small)
  model_matrix_b_samp_small <- model.matrix(formula_b, model_frame)
  
  # efficiencies (alpha)
  formula_a <- community ~ Cycles -1
  model_frame <- model.frame(formula_a, sample_data)
  model_vector_a_samp <- model.matrix(formula_a, model_frame) %>% as.numeric()
  model_frame <- model.frame(formula_a, sample_data_small)
  model_vector_a_samp_small <- model.matrix(formula_a, model_frame) %>% as.numeric()
  
  #counters 
  N_obs_samp_small <- nrow(model_matrix_b_samp_small)
  N_obs_samp <- nrow(sample_data)
  N_b_samp_col <- ncol(model_matrix_b_samp)  
  

stan_data <- list(
  N_species = N_species,   # Number of species in data
  N_obs_samp = N_obs_samp, # Number of observed samples 
  N_obs_mock = N_obs_mock, # Number of observed mock samples
  N_obs_samp_small = N_obs_samp_small, # Number of observed samples 
  N_obs_mock_small = N_obs_mock_small, # Number of observed mock samples
  
  # Observed data of community matrices
  sample_data = sample_data %>% select(-community,-Cycles,-tech_rep),
  mock_data   = mock_data  %>% select(-community,-Cycles,-tech_rep),
  
  # True proportions for mock community
  #mock_true_prop = p_mock_all %>% dplyr::select(contains("sp")),
  alr_mock_true_prop = alr_mock_true_prop,
  alr_mock_true_prop_small = alr_mock_true_prop_small,
  
  # vectors of PCR numbers
  N_pcr_samp = N_pcr_samp,
  N_pcr_mock = N_pcr_mock,
  
  # Design matrices: field samples
  N_b_samp_col = N_b_samp_col,
  model_matrix_b_samp = model_matrix_b_samp,
  model_matrix_b_samp_small = model_matrix_b_samp_small,
  model_vector_a_samp = model_vector_a_samp,
  model_vector_a_samp_small = as.array(model_vector_a_samp_small),
  
  # Design matrices: mock community samples
  N_b_mock_col = N_b_mock_col,
  model_matrix_b_mock = model_matrix_b_mock,
  model_matrix_b_mock_small = model_matrix_b_mock_small,
  model_vector_a_mock = model_vector_a_mock,
  model_vector_a_mock_small = model_vector_a_mock_small,
  
  # Priors
  alpha_prior = c(0,0.1),  # normal prior
  beta_prior = c(0,10),    # normal prior
  tau_prior = c(1.5,1.5)   # gamma prior
)


stan_pars <- c(
  "alpha",
  "beta",
  "eta_samp",
  "eta_mock",
  "tau",
  "mu_samp",
  "mu_mock",
  "int_samp_small"
)

stan_init_f2 <- function(n.chain,N_species){#J_seb,K_seb){
  A <- list()
  for(i in 1:n.chain){
    A[[i]] <- list(
      # tau = runif(N_species-1,0.1,0.5),
      alpha_raw = runif(N_species-1,-0.5,0.5)
    )
  }
  return(A)
}
```

# These are the parameters that are to be monitored during optimization or MCMC:
stan_pars <- c(
  "alpha", # efficiencies relative to the reference species
  "beta",  # parameters for each site (NOT )
  "eta_samp", # overdispersion random effects for each species-site combination (field samples)
  "eta_mock", # overdispersion random effects for each species-site combination (mock samples)
  "tau", # sd of random effects (one for each species (less the reference species))
  "mu_samp", # Predicted proportions for each species-site (unknown samples)
  "mu_mock", # Predicted proportions for each species-site (mock samples)  
  "int_samp_small" # this is the predicted intercept for each site 
)


```{r}
#########################################
#########################################
#Bayesian Estimation
N_CHAIN = 3
Warm = 1000
Iter = 1500
Treedepth = 15
Adapt_delta = 0.70

rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())

## have to change wd here
setwd("~/nwfsc-cb-Toward_Quantitative_Metabarcoding-fba0076/code/Core_Stan_models")

stanMod = stan(file = "quant_metabar_multinom.stan" ,data = stan_data,
               verbose = FALSE, chains = N_CHAIN, thin = 1,
               warmup = Warm, iter = Warm + Iter,
               control = list(adapt_init_buffer = 175,
                              max_treedepth=Treedepth,
                              stepsize=0.01,
                              adapt_delta=Adapt_delta,
                              metric="diag_e"),
               pars = stan_pars,
               refresh = 10,
               boost_lib = NULL,
               init = stan_init_f2(n.chain=N_CHAIN,N_species=N_species),
               sample_file = paste0("./tmpF.csv")
)

pars <- rstan::extract(stanMod, permuted = TRUE)
samp_params <- get_sampler_params(stanMod)

stanMod_summary <- list()
stanMod_summary[["alpha"]] <- summary(stanMod,pars="alpha")$summary
stanMod_summary[["tau"]] <- summary(stanMod,pars=c("tau"))$summary
stanMod_summary[["beta"]] <- summary(stanMod,pars="beta")$summary
stanMod_summary[["eta_samp_raw"]] <- summary(stanMod,pars="eta_samp")$summary
stanMod_summary[["eta_mock_raw"]] <- summary(stanMod,pars="eta_mock")$summary
stanMod_summary[["mu_samp"]] <- summary(stanMod,pars="mu_samp")$summary
stanMod_summary[["mu_mock"]] <- summary(stanMod,pars="mu_mock")$summary
stanMod_summary[["int_samp_small"]] <- summary(stanMod,pars="int_samp_small")$summary

Output <- list(
  mockcom_data = mockcom_data, # raw data from all observations, all communities.
  env = env,  #environmental data
  mc = mc, #mock data
  Species = unique(mc$Species),
  
  # Realizations of simulations (input data)
  p_true = p_mock,
  p_samp_all = sample_data,
  p_mock_all = mock_data,
  
  # stan input objects
  stan_data = stan_data,
  Warm=Warm,
  Iter=Iter,
  
  # Fitted Objects
  stanMod = stanMod, # Full stan Model fitted object
  pars = pars, # MCMC output
  samp_params=samp_params, # Sampler information
  stanMod_summary = stanMod_summary # posterior summaries.
)
```

```{r}
save(Output,file=paste0("/home/kimberly.ledger/nearshore_eDNA/outputs/samples_w_mocks_updated/field_quant/mock_20241122",".Rdata"))
```
