==> code/8_fieldsample_summary.Rmd <==
---
title: "all taxa in fieldsamples"
output: html_document
date: "2024-11-13"
---

this script looks at all taxa in the nearshore eDNA dataset (i.e. not just the species in the mock communities)

load libraries
```{r, message=FALSE}
library(tidyverse)
library(dplyr)
rename <- dplyr::rename
```

load decontaminated reads and filter to retain just the field samples 
```{r}
decontaminated_table <- read.csv("/home/kimberly.ledger/nearshore_eDNA/outputs/decontaminated_asv_table.csv") %>%
  select(!X) %>%
  filter(reads > 0)
```

load sample metadata
```{r}
metadata <- read.csv("/home/kimberly.ledger/nearshore_eDNA/data/nearshore_metadata.csv") %>%
  rename(Sample_ID = sample_ID)
```

assign asv's to taxons 
```{r}
taxonomy <- read.csv("/home/kimberly.ledger/nearshore_eDNA/outputs/taxonomy_20250110_collapsed.csv") %>%
  select(!X) %>%
  select(qseqid, taxon) %>%
  unique() %>%
  rename(ASV = qseqid) 
```

join to decontaminated ASV table
```{r}
decon_table_w_tax <-  decontaminated_table %>%
  left_join(taxonomy, by = "ASV") #%>%
   #filter(!str_starts(Sample_ID, "COM"))
```

```{r}
n_distinct(decon_table_w_tax$ASV)
n_distinct(decon_table_w_tax$taxon)
decon_table_w_tax %>%
  group_by(ASV) %>%
  summarize(total_reads = sum(reads)) %>%
  arrange(total_reads)
sum(decon_table_w_tax$reads)

sample_summary <- decon_table_w_tax %>%
  group_by(Sample_ID, sample_type) %>%
  summarize(total_reads = sum(reads),
            n_ASV = n_distinct(ASV),
            n_tax = n_distinct(taxon))

sample_summary_all <- metadata %>%
  select(Sample_ID) %>%
  left_join(sample_summary)

#write.csv(sample_summary_all, "/home/kimberly.ledger/nearshore_eDNA/outputs/sample_summary.csv")
```



make a summary table of how many reads are assigned to each taxa 
```{r}
taxon_read_summary <- decon_table_w_tax %>%
  group_by(taxon, sample_type) %>%
  summarise(total_reads = sum(reads)) %>%
  pivot_wider(names_from = sample_type, values_from = total_reads, values_fill = 0)
```

convert to taxon table 
```{r}
taxon_table <- decon_table_w_tax %>%
  select(Sample_ID, sample_type, taxon, reads) %>%
  mutate(reads = as.integer(reads)) %>%
  pivot_wider(names_from = taxon, values_from = reads, values_fn = sum, values_fill = 0)
```

calculate proportion of reads per taxa in each technical/PCR replicate 
```{r}
taxon_proportions <- taxon_table %>%
  pivot_longer(cols = -c(Sample_ID, sample_type), names_to = "taxon", values_to = "reads") %>%
  group_by(Sample_ID) %>%
  mutate(total_reads = sum(reads),
         prop = reads/total_reads) %>%
  separate(col = Sample_ID, into = c("bottle", "replicate"), sep = "-")
```

calculate simple mean of taxa proportions for a given bottle 
```{r}
bottle_prop <- taxon_proportions %>%
  group_by(bottle, taxon) %>%
  summarize(simple.N = n(),
            simple.mean = mean(prop),
            simple.SD = sd(prop))
```

join back some metadata
```{r}
metadata_mini <- metadata %>%
  rename(bottle = extraction_ID) %>%
  select(!Sample_ID) %>%
  unique()

bottle_prop_w_meta <- bottle_prop %>%
  filter(taxon != "NA") %>%
  left_join(metadata_mini, by = "bottle") %>%
  filter(sample_type != "mock")

bottle_prop_w_meta %>%
  filter(is.na(alternative_ID)) %>%
  filter(sample_type != "mock")
```


export
```{r}
#write.csv(bottle_prop_w_meta, "~/nearshore_eDNA/outputs/all_taxa_simple_20250110.csv")
```
